+++ src/data/compositemetric/ClusterNodeHealthMetric.js
@@ -1,0 +1,9 @@
+import CompositeMetric from './CompositeMetric';
+
+export default class ClusterNodeHealthMetric extends CompositeMetric {
+    constructor(clusterNode) {
+        super('Cluster Node Health');
+
+        
+    }
+}