+++ src/diagnostic/DiagnosticPane.js
@@ -4,6 +4,7 @@ import { Grid } from 'semantic-ui-react';
 import Overview from './Overview';
 import './DiagnosticPane.css';
 import GeneratePackage from './GeneratePackage';
+import ClusterEventLog from './ClusterEventLog';
 import Ping from './Ping';
 import uuid from 'uuid';
 
@@ -22,6 +23,11 @@ class DiagnosticPane extends Component {
                         </Grid.Column>
                     </Grid.Row>
 
+                    {/* 
+                      * Diagnostic functionality only applies for enterprise installs, because of
+                      * the stored procedures they rely on.
+                      */}
+                    { window.halinContext.isEnterprise() ? 
                     <Grid.Row columns={1}>
                         <Grid.Column>
                             <h3>Run Diagnostics</h3> 
@@ -36,7 +42,14 @@ class DiagnosticPane extends Component {
                                 driver={this.props.driver}
                             />
                         </Grid.Column>
-                    </Grid.Row>
+                    </Grid.Row> : '' }
+
+                    { window.halinContext.isCluster() ? 
+                    <Grid.Row columns={1}>
+                        <Grid.Column>
+                            <ClusterEventLog />
+                        </Grid.Column>
+                    </Grid.Row> : '' }
 
                     <Grid.Row columns={1}>
                         <Grid.Column>