+++ src/higherOrderComponents.js
@@ -1,6 +1,32 @@
 import React, { Component } from 'react';
 import { Icon, Message } from 'semantic-ui-react';
 
+const smallCentered = { maxWidth: 300, margin: 'auto' };
+
+const clusterOnlyComponent = (WrappedComponent, heading) => {
+    return class extends Component {
+        render() {
+            const nodes = window.halinContext.clusterNodes;
+
+            if (nodes && nodes.length > 1) {
+                return <WrappedComponent {...this.props} />;
+            }
+
+            return (
+                <div className='ClusterOnly'>
+                    { heading ? <h3>{heading}</h3> : '' }
+                    <Message warning icon style={smallCentered}>
+                        <Icon name='warning' />
+                        <Message.Content >
+                            Only available for Neo4j Clusters
+                        </Message.Content>
+                    </Message>
+                </div>
+            )
+        }
+    }
+};
+
 const adminOnlyComponent = (WrappedComponent, heading) => {
     return class extends Component {
         render() {
@@ -18,7 +44,7 @@ const adminOnlyComponent = (WrappedComponent, heading) => {
             return (
                 <div className='AdminOnly'>
                     { heading ? <h3>{heading}</h3> : '' }
-                    <Message warning icon>
+                    <Message warning icon style={smallCentered}>
                         <Icon name='warning' />
                         <Message.Content>
                             Only users with role 'admin' may use this function.
@@ -41,10 +67,10 @@ const enterpriseOnlyComponent = (WrappedComponent, heading) => {
             return (
                 <div className='EnterpriseOnly'>
                     { heading ? <h3>{heading}</h3> : '' }
-                    <Message warning icon>
+                    <Message warning icon style={smallCentered}>
                         <Icon name='warning' />
                         <Message.Content>
-                            This function is only available in Neo4j Enterprise
+                            Only available in Neo4j Enterprise
                         </Message.Content>
                     </Message>
                 </div>
@@ -56,4 +82,5 @@ const enterpriseOnlyComponent = (WrappedComponent, heading) => {
 export default {
     adminOnlyComponent,
     enterpriseOnlyComponent,
+    clusterOnlyComponent,
 };