+++ src/overview/ClusterOverviewPane.js
@@ -9,12 +9,19 @@ import PageCacheTracking from './PageCacheTracking';
 import TransactionsOpen from './TransactionsOpen';
 import UsedMemory from './UsedMemory';
 import OpenFileDescriptors from './OpenFileDescriptors';
+// import ClusterView from './ClusterView';
 
 export default class ClusterOverviewPane extends Component {
     render() {
         return (
             <div className="PerformancePane">
                 <Grid divided='vertically'>
+                    {/* <Grid.Row columns={1}>
+                        <Grid.Column>
+                            <ClusterView />
+                        </Grid.Column>
+                    </Grid.Row> */}
+
                     <Grid.Row columns={3}>
                         <Grid.Column>
                             <ClusterMemory />